Initial motion detection code
This is a first run at the code that detects changes in
direction. It uses Core Motion and utilizes the lower
power, faster code (because we care less about 'accuracy'
in terms of mapping to real 3D space and more about efficient,
less power-hungry code). It basically takes the rotational velocity
and motion in 3D space (roll, pitch, yaw) and checks for certain
thresholds to determine if a motion is made to turn right or
left or to go forward or reverse.