Refactor socket communication
To Transmit:
The MainViewController calls to the transmit
delegate to start and stop the transmission.
It also sends commands via GCDAsyncSocket
(it uses its own GCD queue). This is all done
by the ArduinoTransmitHandler.

To Receive:
The MainViewController calls to the receive
delegate to start and stop listening. This
is done via GCDAsyncSocket (it uses its
own GCD queue). It also implements the
datasource so that it gets updated by
the handler when it receives data.

I should note that, because I don't have
the Arduino and have to go to the lab,
it's possible I broke like, a lot of shit
when I did this refactoring. Likely the following
commits are fixes patching this work.

- Add ArduinoTransmitHandler (and delegate)
- Add ArduinoReceiveHandler (and delegate and
datasource)
- Update MainViewController
- Add Direction model (really, it's just an
enum I moved to a separate file for easy inclusion)